Removed redundant setting of tempurature to 1 in new_chatgpt(). Fixed the type of temperature (was int before, but should have been float per the openai api).

+++ src/construct_json.h
@@ -12,7 +12,7 @@ https://stackoverflow.com/questions/36998026/append-json-object-arrays-in-c-usin
 #define free_chatgpt(chat) json_object_put(chat)
 
 void add_text_prompt(json_object *input, const char *role, const char *prompt_text);
-void set_temp(json_object *input, int temp);
+void set_temp(json_object *input, float temp);
 void set_model(json_object *input, char* model);
 json_object *new_chatgpt(void);
 //void add_json_prompt(json_object *dest, json_object *src);
